Iris Flower Dataset

The Iris flower dataset or Fisher's Iris dataset is a multivariate dataset introduced by Sir Ronald Fisher in 1936 as an example of discriminant analysis. It includes measurements for 150 iris flowers from three species, with 50 from each species. The measurements include the length and the width of the sepals and petals.

Format

A data frame with 150 rows and 5 variables:

Sepal.Length

Length of the sepal in cm.

Sepal.Width

Width of the sepal in cm.

Petal.Length

Length of the petal in cm.

Petal.Width

Width of the petal in cm.

Type

A factor with levels `setosa`, `versicolor`, and `virginica` indicating the species of each observation.

Source

Fisher, R.A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179-188.